chore: moved store and constants to utils folder
