+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/__tests__/checkoutConfiguration.test.js
@@ -1,7 +1,7 @@
 /**
  * @jest-environment jsdom
  */
-const store = require('../../../../../store');
+const store = require('../../../../../utils/store');
 const helpers = require('../helpers');
 const httpClient = require('../../../js/commons/httpClient');
 const {
