+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/helpers.js
@@ -1,5 +1,5 @@
-const store = require('../../../../store');
-const constants = require('../constants');
+const store = require('../../../../utils/store');
+const constants = require('../../../../utils/constants');
 const { httpClient } = require('../commons/httpClient');
 
 function assignPaymentMethodValue() {
