+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/paymentMethodsConfiguration/giftcards/giftcardsConfig.js
@@ -6,7 +6,11 @@ const {
   attachGiftCardCancelListener,
   createElementsToShowRemainingGiftCardAmount,
 } = require('../../giftcards');
-const { GIFTCARD, SUCCESS, NOTENOUGHBALANCE } = require('../../../constants');
+const {
+  GIFTCARD,
+  SUCCESS,
+  NOTENOUGHBALANCE,
+} = require('../../../../../../utils/constants');
 const { initializeCheckout } = require('../../renderGenericComponent');
 
 class GiftCardConfig {
