+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/renderGenericComponent.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-unsafe-optional-chaining */
-const store = require('../../../../store');
+const store = require('../../../../utils/store');
 const { renderPaymentMethod } = require('./renderPaymentMethod');
 const helpers = require('./helpers');
-const constants = require('../constants');
+const constants = require('../../../../utils/constants');
 
 function setCheckoutConfiguration(checkoutOptions) {
   const setField = (key, val) => val && { [key]: val };
