+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/renderPaymentMethod.js
@@ -1,6 +1,6 @@
-const store = require('../../../../store');
+const store = require('../../../../utils/store');
 const helpers = require('./helpers');
-const constants = require('../constants');
+const constants = require('../../../../utils/constants');
 
 function getFallback(paymentMethod) {
   const fallback = {};
