+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/__tests__/applePayExpress.test.js
@@ -4,11 +4,11 @@
 
 const ApplePay = require('../paymentMethods/applepay/applepay');
 const httpClient = require('../../../js/commons/httpClient');
-const store = require('../../../../../../cartridge/store');
+const store = require('../../../../../../cartridge/utils/store');
 const helpers = require('../../../js/adyen_checkout/helpers');
 
 jest.mock('../../../js/commons/httpClient');
-jest.mock('../../../../../../cartridge/store');
+jest.mock('../../../../../../cartridge/utils/store');
 jest.mock('../../../js/adyen_checkout/helpers');
 jest.mock('../initializeCheckout');
 jest.mock('../../commons');
