+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/paymentMethods/amazonpay/amazonPayExpressPart1.js
@@ -1,5 +1,5 @@
-const store = require('../../../../../../store');
-const { AMAZON_PAY } = require('../../../constants');
+const store = require('../../../../../../utils/store');
+const { AMAZON_PAY } = require('../../../../../../utils/constants');
 const { initializeCheckout } = require('../../initializeCheckout');
 
 class AmazonPay {
