+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/commons/index.js
@@ -1,5 +1,10 @@
-const store = require('../../../../store');
-const { PAYPAL, APPLE_PAY, AMAZON_PAY, GOOGLE_PAY } = require('../constants');
+const store = require('../../../../utils/store');
+const {
+  PAYPAL,
+  APPLE_PAY,
+  AMAZON_PAY,
+  GOOGLE_PAY,
+} = require('../../../../utils/constants');
 const { httpClient } = require('./httpClient');
 
 module.exports.onFieldValid = function onFieldValid(data) {
