+++ src/cartridges/app_adyen_SFRA/cartridge/utils/constants.js
